[ghstack-poisoned]

+++ tests/unit_tests/test_permute_indices_kernel.py
@@ -24,7 +24,7 @@ def fill_indices_cpu(
     """CPU implementation for reference."""
     # We need to preallocate the output
     device = tokens_per_expert_group.device
-    permuted_indices = torch.full((max_len,), -1, dtype=torch.int32, device=device)
+    permuted_indices = torch.full((max_len,), -1, dtype=torch.int64, device=device)
     # Fill the permuted indices
     # For each local expert
     for e in range(experts_per_rank):
@@ -42,7 +42,7 @@ def fill_indices_cpu(
                     permuted_indices[write_start:end_idx] = torch.arange(
                         start_index,
                         start_index + (end_idx - write_start),
-                        dtype=torch.int32,
+                        dtype=torch.int64,
                         device=device,
                     )
             write_start += length
